feat: rename upgrade, add event
