+++ app/upgrades/v5beta2/constants.go
@@ -1,11 +1,11 @@
-package v6
+package v5beta2
 
 import (
 	storetypes "cosmossdk.io/store/types"
 	"github.com/Zenrock-Foundation/zrchain/v5/app/upgrades"
 )
 
-const UpgradeName = "v6"
+const UpgradeName = "v5beta2"
 
 var Upgrade = upgrades.Upgrade{
 	UpgradeName:          UpgradeName,
